Promote CSV upload preview component
