+++ vuu-ui/packages/vuu-table-extras/src/csv-upload/README.md
@@ -97,7 +97,7 @@ If the user cancels at any point:
 | Callback | Phase transition | Notes |
 |---|---|---|
 | `onProcessingStarted` | `→ processing` | Fires before parsing begins. No data available yet. |
-| `onImportSessionStarted` | `→ preview-ready` | Provides a session `DataSource`. Extract `dataSource.table` (`CsvUploadSessionTable`) and pass it to `useCsvUploadSessionPreview` to build a preview UI without mutating the shared session datasource. |
+| `onImportSessionStarted` | `→ preview-ready` | Provides the populated session `DataSource`. |
 | `onImportSessionEnded` | `→ imported` or `→ idle` | `reason` is `"saved"` on successful import, `"discarded"` on cancel, `"failed"` on error. `sessionTable` contains the Vuu session table reference. |
 | `onImported` | `→ imported` | Provides normalized `tableData`. |
 | `onError` | `→ failed` | See [Error Types](#error-types) below. |
@@ -163,64 +163,6 @@ type CsvErrorMap<TError extends string> = {
 | `MAX_ROWS_EXCEEDED` | The CSV contains more rows than the `maxRows` limit. Reported as a `fileError`. |
 | `EMPTY_NON_STRING_VALUE` | A non-string column cell is empty. Reported as a `rowError`. |
 | `TYPE_MISMATCH` | A cell value cannot be coerced to the column's server data type. Reported as a `rowError`. |
-
-### Session preview
-
-When `onImportSessionStarted` fires, the edit session table is populated with one row per uploaded CSV row. Each row has the source table's columns plus a `vuuMsg` column. For rows with validation errors, `vuuMsg` contains a human-readable summary including the CSV row number:
-
-```
-"Row 3: price: Value 'abc' is not a valid double; ric: Empty value is not allowed for non-string columns."
-```
-
-Rows that passed validation have an empty `vuuMsg`.
-
-The recommended way to build a preview UI is the `useCsvUploadSessionPreview` hook, which fetches the session table schema and creates a dedicated datasource — keeping the shared session datasource untouched:
-
-```tsx
-import {
-  CsvUpload,
-  type CsvUploadSessionTable,
-  useCsvUploadSessionPreview,
-} from "@vuu-ui/vuu-table-extras";
-
-const ErrorsTable = ({ sessionTable }: { sessionTable: CsvUploadSessionTable | undefined }) => {
-  const { isLoadingPreview, previewDataSource, previewError } =
-    useCsvUploadSessionPreview(sessionTable);
-
-  useEffect(() => {
-    if (previewDataSource) {
-      // Show only rows that have errors
-      previewDataSource.filter = { filter: 'vuuMsg > ""' };
-    }
-  }, [previewDataSource]);
-
-  if (!previewDataSource || isLoadingPreview) return null;
-  if (previewError) return <div>{previewError}</div>;
-  return <Table dataSource={previewDataSource} config={errorTableConfig} ... />;
-};
-
-const MyApp = () => {
-  const [sessionTable, setSessionTable] = useState<CsvUploadSessionTable | undefined>();
-
-  return (
-    <CsvUpload
-      dataSource={dataSource}
-      onImportSessionStarted={(sessionDs) => setSessionTable(sessionDs.table as CsvUploadSessionTable)}
-      onImportSessionEnded={() => setSessionTable(undefined)}
-    >
-      <ErrorsTable sessionTable={sessionTable} />
-    </CsvUpload>
-  );
-};
-```
-
-The `useCsvUploadSessionPreview` hook:
-- Calls `getServerAPI().getTableSchema(sessionTable)` to fetch the full schema
-- Builds `ColumnDescriptor[]` from the schema, giving the `vuuMsg` column a wider default width and the label `"Error"`
-- Creates a dedicated `VuuDataSource` for the session table
-- Cleans up when `sessionTable` becomes `undefined`
-
-To show only rows with errors, apply the filter `'vuuMsg > ""'` on the `previewDataSource` (not on the datasource from `onImportSessionStarted`).
 
 ## Parse Options
