+++ vuu-ui/packages/vuu-table-extras/src/csv-upload/data-upload-preview/DataUploadPreview.css
@@ -1,0 +1,10 @@
+.vuuDataUploadPreview {
+  display: flex;
+  flex-direction: column;
+  height: 420px;
+  width: 900px;
+
+  .vuuDataUploadPreview-content {
+    flex: 1 1 auto;
+  }
+}
