+++ vuu-ui/packages/vuu-table/src/__tests__/__component__/Table-editing.playwright.test.tsx
@@ -6,7 +6,6 @@ import {
   EditableInstrumentsInlineEdit,
   EditableInstrumentsWithInlineAddRow,
   TestTableEmpty,
-  TestTableEmptyWithUpload,
   TestTableFIveRows,
   TwoEditableInstruments,
 } from "../../../../../showcase/src/examples/Table/Editing.examples";
@@ -104,90 +103,6 @@ test.describe("Inline add row", () => {
 });
 
 test.describe("Test table editing", () => {
-  test("uploads CSV rows directly from the toolbar", async ({
-    mount,
-    page,
-  }) => {
-    await mount(<TestTableEmptyWithUpload />);
-
-    await expect(
-      page.getByRole("button", { name: "Upload (preview)" }),
-    ).toBeVisible();
-    await page.getByRole("button", { name: "Upload (direct)" }).click();
-    await page.locator('input[type="file"]').setInputFiles({
-      name: "test-edit.csv",
-      mimeType: "text/csv",
-      buffer: Buffer.from(
-        "id,description,quantity,price,enabled,externalId\nCSV-001,Direct upload,10,12.5,true,1001\n",
-      ),
-    });
-    await page.getByRole("button", { name: "Import", exact: true }).click();
-
-    await expect(
-      page.getByRole("heading", { name: "Upload Data" }),
-    ).not.toBeVisible();
-    await expect(page.locator(".vuuDatasourceStats-value").last()).toHaveText(
-      "1",
-    );
-  });
-
-  test("uploads CSV rows into an editable session", async ({ mount, page }) => {
-    await mount(<TestTableEmptyWithUpload />);
-
-    await page.getByRole("button", { name: "Upload (preview)" }).click();
-    await expect(
-      page.getByRole("heading", { name: "Upload Data" }),
-    ).toBeVisible();
-
-    await page.locator('input[type="file"]').setInputFiles({
-      name: "test-edit.csv",
-      mimeType: "text/csv",
-      buffer: Buffer.from(
-        "id,description,quantity,price,enabled,externalId\nCSV-001,Uploaded row,10,12.5,true,1001\nCSV-002,Delete me,20,25,true,1002\n",
-      ),
-    });
-    await page.getByRole("button", { name: "Import", exact: true }).click();
-
-    await expect(
-      page.getByRole("heading", { name: "Edit uploaded data" }),
-    ).toBeVisible();
-    await expect(page.locator('input[value="CSV-001"]')).toBeVisible();
-    await expect(page.locator(".vuuInlineAddRow")).toHaveCount(0);
-
-    const description = page
-      .getByRole("textbox", {
-        name: "description",
-        exact: true,
-      })
-      .first();
-    await description.fill("Edited upload");
-    await description.press("Enter");
-    await expect(description).toHaveValue("Edited upload");
-
-    const deletedRow = page
-      .locator('input[value="CSV-002"]')
-      .locator('xpath=ancestor::*[@role="row"]');
-    await deletedRow
-      .getByRole("checkbox", { name: "Press space to select row" })
-      .click();
-    await page.getByRole("button", { name: "Delete", exact: true }).click();
-    await expect(deletedRow).toContainClass("vuuTableRow-deleted");
-
-    const submitButton = page.getByRole("button", {
-      name: "Submit",
-      exact: true,
-    });
-    await expect(submitButton).toBeEnabled();
-    await submitButton.click();
-
-    await expect(
-      page.getByRole("heading", { name: "Edit uploaded data" }),
-    ).not.toBeVisible();
-    await expect(page.locator(".vuuDatasourceStats-value").last()).toHaveText(
-      "1",
-    );
-  });
-
   test("inserts a row with an unchecked boolean without validation warnings", async ({
     mount,
     page,
